+++ CHANGELOG.md
@@ -14,6 +14,13 @@ All notable changes to this project will be documented in this file.
 
 - Option title for restricted list renamed.
 
+### Change :
+
+- **Customs Hooks: (read and write in storage)**
+  - Access to restrict list is now centralise in the useRestrictList hook.
+  - Access to active sessions is now centralise in the useActiveSessions hook.
+  - Access to settings variable is now centralise in the useSettings hook.
+
 ## v1.0.0
 
 ### Features :